Create problem 3
project euler problem 3

The prime factors of 13195 are 5, 7, 13 and 29.

What is the largest prime factor of the number 600851475143 ?

code produces correct answer of 6857. however, the code should return 6857, but produces non-prime divisors after 6857. added System.out.println(i) in order to see the values of the largestfactor method. needs further investigation.